Refactor toom-cook-6-5
